Make joining the shared tmux session easier for both parties
This change ensures that when the invitee connects to the machine, they
are automatically attached to the tmux session. Likewise, the inviter is
automatically joined to the same shared tmux session provided that they
are not already in it. Another check is added for the case where the
inviter is already in a tmux session but not the shared one. In that
case, the script tells them the exact command to run.